drf_login_api/requirements.txt
Django
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0